requirements.txt
aiohttp==3.7.4.post0
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
chardet==4.0.0
Django==3.2.7
django-cors-headers==3.10.0
django-filter==2.4.0
django-graphql-jwt==0.3.4
djongo==1.3.6
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
multidict==5.2.0
Pillow==8.4.0
promise==2.3
psycopg2-binary==2.9.1
PyJWT==2.1.0
pymongo==3.12.0
pytz==2021.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.2.4
text-unidecode==1.3
typing-extensions==3.10.0.2
yarl==1.7.0
django-background-tasks==1.2.5
git+git://github.com/altair-viz/altair.git
matplotlib==3.5.0
scikit-learn==1.0.1
tensorflow==2.7.0
